Modify Setcontent to match current repo (without Runtime)
